# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Circle": {
        "properties": {
          "height": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "The height (used by cylinder/cone volume and surface area).",
            "title": "Height"
          },
          "radius": {
            "description": "The radius of the circle.",
            "title": "Radius",
            "type": "number"
          }
        },
        "required": [
          "radius"
        ],
        "title": "Circle",
        "type": "object"
      },
...
      "Ellipse": {
        "properties": {
          "axis3": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "The third semi-axis (required for ellipsoid volume/surface area).",
            "title": "Axis3"
          },
          "semi_major": {
            "description": "The semi-major axis of the ellipse.",
            "title": "Semi Major",
            "type": "number"
          },
          "semi_minor": {
            "description": "The semi-minor axis of the ellipse.",
            "title": "Semi Minor",
            "type": "number"
          }
        },
        "required": [
          "semi_major",
          "semi_minor"
        ],
        "title": "Ellipse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Parallelogram": {
        "properties": {
          "side1": {
            "description": "The length of the first side.",
            "title": "Side1",
            "type": "number"
          },
          "side2": {
            "description": "The length of the second side.",
            "title": "Side2",
            "type": "number"
          },
          "theta": {
            "description": "The angle (in degrees) between the two sides.",
            "title": "Theta",
            "type": "number"
          }
        },
        "required": [
          "side1",
          "side2",
          "theta"
        ],
        "title": "Parallelogram",
        "type": "object"
      },
...
      "SemiCircle": {
        "properties": {
          "radius": {
            "description": "The radius of the semicircle.",
            "title": "Radius",
            "type": "number"
          }
        },
        "required": [
          "radius"
        ],
        "title": "SemiCircle",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "The Elephant Calculator API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/geometry/area/circle": {
      "post": {
        "description": "Calculate the area of a circle.\n\nArgs:\n    data (Circle): Contains the radius of the circle.\n\nReturns:\n    float: The area of the circle.",
        "operationId": "calculate_area_circle_api_geometry_area_circle_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Circle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Calculate Area Circle Api Geometry Area Circle Post",
                  "type": "number"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calculate Area Circle",
        "tags": [
          "Geometry"
        ]
      }
    },
    "/api/geometry/area/ellipse": {
      "post": {
        "description": "Calculate the area of an ellipse.\n\nArgs:\n    data (Ellipse): Contains the semi-major and semi-minor axes of the ellipse.\n\nReturns:\n    float: The area of the ellipse.",
        "operationId": "calculate_area_ellipse_api_geometry_area_ellipse_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Ellipse"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Calculate Area Ellipse Api Geometry Area Ellipse Post",
                  "type": "number"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calculate Area Ellipse",
        "tags": [
          "Geometry"
        ]
      }
    },
    "/api/geometry/area/parallelogram": {
      "post": {
        "description": "Calculate the area of a parallelogram.\n\nArgs:\n    data (Parallelogram): Contains the two side lengths and the angle between them (in degrees).\n\nReturns:\n    float: The area of the parallelogram.",
        "operationId": "calculate_area_parallelogram_api_geometry_area_parallelogram_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Parallelogram"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Calculate Area Parallelogram Api Geometry Area Parallelogram Post",
                  "type": "number"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calculate Area Parallelogram",
        "tags": [
          "Geometry"
        ]
      }
    },
    "/api/geometry/area/polygon": {
      "post": {
        "description": "Calculate the area of a regular polygon.\n\nArgs:\n    data (Polygon): Contains the number of sides and the length of each side.\n\nReturns:\n    float: The area of the polygon.",
        "operationId": "calculate_area_polygon_api_geometry_area_polygon_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Polygon"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Calculate Area Polygon Api Geometry Area Polygon Post",
                  "type": "number"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calculate Area Polygon",
        "tags": [
          "Geometry"
        ]
      }
    },
    "/api/geometry/area/rectangle": {
      "post": {
        "description": "Calculate the area of a rectangle.\n\nArgs:\n    data (Rectangle): Contains the length and width of the rectangle.\n\nReturns:\n    float: The area of the rectangle.",
        "operationId": "calculate_area_rectangle_api_geometry_area_rectangle_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Rectangle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Calculate Area Rectangle Api Geometry Area Rectangle Post",
                  "type": "number"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calculate Area Rectangle",
        "tags": [
          "Geometry"
        ]
      }
    },
    "/api/geometry/area/rhombus": {
      "post": {
        "description": "Calculate the area of a rhombus.\n\nArgs:\n    data (Rhombus): Contains the lengths of the diagonals of the rhombus.\n\nReturns:\n    float: The area of the rhombus.",
        "operationId": "calculate_area_rhombus_api_geometry_area_rhombus_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Rhombus"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Calculate Area Rhombus Api Geometry Area Rhombus Post",
                  "type": "number"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calculate Area Rhombus",
        "tags": [
          "Geometry"
        ]
      }
    },
    "/api/geometry/area/semi_circle": {
      "post": {
        "description": "Calculate the area of a semicircle.\n\nArgs:\n    data (SemiCircle): Contains the radius of the semicircle.\n\nReturns:\n    float: The area of the semicircle.",
        "operationId": "calculate_area_semi_circle_api_geometry_area_semi_circle_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SemiCircle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Calculate Area Semi Circle Api Geometry Area Semi Circle Post",
                  "type": "number"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calculate Area Semi Circle",
        "tags": [
          "Geometry"
        ]
      }
    },
    "/api/geometry/area/square": {
      "post": {
        "description": "Calculate the area of a square.\n\nArgs:\n    data (Square): Contains the side length of the square.\n\nReturns:\n    float: The area of the square.",
        "operationId": "calculate_area_square_api_geometry_area_square_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Square"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Calculate Area Square Api Geometry Area Square Post",
                  "type": "number"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calculate Area Square",
        "tags": [
          "Geometry"
        ]
      }
    },
    "/api/geometry/area/trapezoid": {
      "post": {
        "description": "Calculate the area of a trapezoid.\n\nArgs:\n    data (Trapezoid): Contains the lengths of the two bases and the height of the trapezoid.\n\nReturns:\n    float: The area of the trapezoid.",
        "operationId": "calculate_area_trapezoid_api_geometry_area_trapezoid_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Trapezoid"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Calculate Area Trapezoid Api Geometry Area Trapezoid Post",
                  "type": "number"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calculate Area Trapezoid",
        "tags": [
          "Geometry"
        ]
      }
    },
    "/api/geometry/area/triangle/base_height": {
      "post": {
        "description": "Calculate the area of a triangle using its base and height.\n\nArgs:\n    data (TriangleBaseHeight): Contains the base length and height of the triangle.\n\nReturns:\n    float: The area of the triangle.",
        "operationId": "calculate_area_triangle_base_height_api_geometry_area_triangle_base_height_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriangleBaseHeight"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Calculate Area Triangle Base Height Api Geometry Area Triangle Base Height Post",
                  "type": "number"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calculate Area Triangle Base Height",
        "tags": [
          "Geometry"
        ]
      }
    },
...
  }
}